+++ test/server/admin/admin_test.cc
@@ -225,35 +225,6 @@ TEST_P(AdminInstanceTest, RejectHandlerWithEmbeddedQuery) {
                                                      callback, true, false)));
 }
 
-TEST_P(AdminInstanceTest, EscapeHelpTextWithPunctuation) {
-  auto callback = [](absl::string_view, Http::HeaderMap&, Buffer::Instance&,
-                     AdminStream&) -> Http::Code { return Http::Code::Accepted; };
-
-  // It's OK to have help text with HTML characters in it, but when we render the home
-  // page they need to be escaped.
-  const std::string planets = "jupiter>saturn>mars";
-  EXPECT_TRUE(admin_.addHandler("/planets", planets, callback, true, false));
-
-  Http::TestResponseHeaderMapImpl header_map;
-  Buffer::OwnedImpl response;
-  EXPECT_EQ(Http::Code::OK, getCallback("/", header_map, response));
-  const Http::HeaderString& content_type = header_map.ContentType()->value();
-  EXPECT_THAT(std::string(content_type.getStringView()), HasSubstr("text/html"));
-  EXPECT_EQ(-1, response.search(planets.data(), planets.size(), 0, 0));
-  const std::string escaped_planets = "jupiter&gt;saturn&gt;mars";
-  EXPECT_NE(-1, response.search(escaped_planets.data(), escaped_planets.size(), 0, 0));
-}
-
-TEST_P(AdminInstanceTest, HelpUsesFormForMutations) {
-  Http::TestResponseHeaderMapImpl header_map;
-  Buffer::OwnedImpl response;
-  EXPECT_EQ(Http::Code::OK, getCallback("/", header_map, response));
-  const std::string logging_action = "<form action='logging' method='post'";
-  const std::string stats_href = "<a href='stats'";
-  EXPECT_NE(-1, response.search(logging_action.data(), logging_action.size(), 0, 0));
-  EXPECT_NE(-1, response.search(stats_href.data(), stats_href.size(), 0, 0));
-}
-
 class AdminTestingPeer {
 public:
   AdminTestingPeer(AdminImpl& admin)
